+++ src/main/java/de/adesso/kicker/match/controller/MatchController.java
@@ -5,7 +5,6 @@ import de.adesso.kicker.match.exception.InvalidCreatorException;
 import de.adesso.kicker.match.exception.SamePlayerException;
 import de.adesso.kicker.match.persistence.Match;
 import de.adesso.kicker.match.service.MatchService;
-import de.adesso.kicker.notification.service.NotificationService;
 import de.adesso.kicker.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.MailException;
@@ -26,8 +25,6 @@ public class MatchController {
     private final MatchService matchService;
 
     private final UserService userService;
-
-    private final NotificationService notificationService;
 
     @GetMapping("/add")
     public ModelAndView getAddMatch() {
@@ -69,12 +66,10 @@ public class MatchController {
     private ModelAndView defaultAddMatchView(ModelAndView modelAndView) {
         var match = new Match();
         var user = userService.getLoggedInUser();
-        var notifications = notificationService.getNotificationsByReceiver(user);
         var users = userService.getAllUsers();
         modelAndView.addObject("match", match);
         modelAndView.addObject("currentUser", user);
         modelAndView.addObject("users", users);
-        modelAndView.addObject("notifications", notifications);
         modelAndView.setViewName("sites/matchresult.html");
         return modelAndView;
     }
