+++ src/main/java/de/adesso/kicker/match/service/events/MatchVerifiedEvent.java
@@ -1,15 +1,15 @@
-package de.adesso.kicker.events.match;
+package de.adesso.kicker.match.service.events;
 
 import de.adesso.kicker.match.persistence.Match;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
 
 @Getter
-public class MatchDeclinedEvent extends ApplicationEvent {
+public class MatchVerifiedEvent extends ApplicationEvent {
 
     private final Match match;
 
-    public MatchDeclinedEvent(Object source, Match match) {
+    public MatchVerifiedEvent(Object source, Match match) {
         super(source);
         this.match = match;
     }
