+++ src/main/java/de/adesso/kicker/user/persistence/RankingRepository.java
@@ -1,4 +1,4 @@
-package de.adesso.kicker.statistics.ranking.persistence;
+package de.adesso.kicker.user.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
