+++ src/main/java/de/adesso/kicker/user/persistence/User.java
@@ -1,6 +1,5 @@
 package de.adesso.kicker.user.persistence;
 
-import de.adesso.kicker.statistics.ranking.persistence.Ranking;
 import lombok.Data;
 import lombok.NoArgsConstructor;
