+++ src/test/java/de/adesso/kicker/match/MatchControllerIntegrationTest.java
@@ -1,9 +1,9 @@
 package de.adesso.kicker.match;
 
 import de.adesso.kicker.email.EmailService;
-import de.adesso.kicker.events.match.MatchVerificationSentEvent;
 import de.adesso.kicker.match.persistence.MatchRepository;
 import de.adesso.kicker.notification.matchverificationrequest.persistence.MatchVerificationRequestRepository;
+import de.adesso.kicker.notification.matchverificationrequest.service.events.MatchVerificationSentEvent;
 import de.adesso.kicker.user.persistence.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
